all: fix retrying of forward tunnels when SSH connection dies
When operating in forward tunnel mode, the main loop in bindTunnel
has no feedback mechanism to cancel all subsequent routines when the
underlying SSH tunnel dies.

The fix for this is simple. We start a goroutine that Waits on the
SSH connection. If this ever returns from Wait, then we know that
the SSH connection died for some reason. Thus, we can invoke the
cancelation sequence.