Add monitor for dead TCP connections
When the underlying TCP connection abruptly dies, the SSH session can hang
indefinitely in a Read call. In order to actually detect a dead session due
to dead TCP connections, periodically send a message to the server to
invoke a response. If no response is received after some number of tries,
close the SSH session so that it can be retried.