Add configuration option for known hosts file
A recent (and breaking) change to the ssh package requires clients to
specify a callback to validate the remote host. Prior to the change,
the default (insecure) behavior was to accept any remote host.
This default was insecure in that it allowed for MITM attacks.
The new behavior requires the user to specify a host key callback,
which we implement by reading a set of known_hosts files provided
in the KnownHostFiles field in the config.

See https://golang.org/issue/19767